+++ src/config.ts
@@ -19,7 +19,7 @@ import {
 	getOracleTree,
 	registerOracleTree
 } from './module/features/customoracles'
-import { OracleTable } from './module/roll-table/roll-table'
+import { OracleTable } from './module/roll-table/oracle-table'
 
 export interface EmitterEvents extends Record<EventType, unknown> {
 	highlightMove: string // Foundry UUID
