+++ src/module/actor/actor.ts
@@ -4,7 +4,7 @@ import type {
 	SiteDataPropertiesData
 } from './actortypes'
 import type { SFCharacterMoveSheet } from './sheets/sf-charactermovesheet'
-import { OracleTable } from '../roll-table/roll-table'
+import { OracleTable } from '../roll-table/oracle-table'
 
 let CREATE_DIALOG: CreateActorDialog
