+++ src/module/dataforged/import.ts
@@ -20,8 +20,8 @@ import { IronswornActor } from '../actor/actor'
 import type { IronswornItem } from '../item/item'
 import { IronswornJournalEntry } from '../journal/journal-entry'
 import { IronswornJournalPage } from '../journal/journal-entry-page'
-import { OracleTable } from '../roll-table/roll-table'
-import { OracleTableResult } from '../roll-table/table-result'
+import { OracleTable } from '../roll-table/oracle-table'
+import { OracleTableResult } from '../roll-table/oracle-table-result'
 import {
 	ISAssetTypes,
 	ISMoveCategories,
