+++ src/module/features/pack-cache.ts
@@ -2,7 +2,7 @@ import type { BaseAdventure } from '@league-of-foundry-developers/foundry-vtt-ty
 import type { IronswornActor } from '../actor/actor'
 import type { IronswornItem } from '../item/item'
 import { IronswornJournalEntry } from '../journal/journal-entry'
-import type { OracleTable } from '../roll-table/roll-table'
+import type { OracleTable } from '../roll-table/oracle-table'
 
 const ONE_MINUTE_IN_MS = 60 * 1000
